+++ hc/api/tests/test_notify_matrix.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from datetime import timedelta as td
 from unittest.mock import Mock, patch
+from urllib.parse import quote
 
 from django.test.utils import override_settings
 from django.utils.timezone import now
@@ -12,7 +13,10 @@ from hc.api.models import Channel, Check, Notification
 from hc.test import BaseTestCase
 
 
-class NotifyPagertreeTestCase(BaseTestCase):
+@override_settings(
+    MATRIX_HOMESERVER="https://example.net", MATRIX_ACCESS_TOKEN="test-token"
+)
+class NotifyMatrixTestCase(BaseTestCase):
     def setUp(self) -> None:
         super().setUp()
 
@@ -23,8 +27,8 @@ class NotifyPagertreeTestCase(BaseTestCase):
         self.check.save()
 
         self.channel = Channel(project=self.project)
-        self.channel.kind = "pagertree"
-        self.channel.value = "123"
+        self.channel.kind = "matrix"
+        self.channel.value = "!foo:example.org"
         self.channel.save()
         self.channel.checks.add(self.check)
 
@@ -35,37 +39,23 @@ class NotifyPagertreeTestCase(BaseTestCase):
         self.channel.notify(self.check)
         assert Notification.objects.count() == 1
 
-        payload = mock_post.call_args.kwargs["json"]
-        self.assertEqual(payload["event_type"], "trigger")
-        self.assertIn("Foo is DOWN.", payload["description"])
-        self.assertIn("Last ping was an hour ago.", payload["description"])
-
-    @override_settings(PAGERTREE_ENABLED=False)
-    def test_it_requires_pagertree_enabled(self) -> None:
-        self.channel.notify(self.check)
-
-        n = Notification.objects.get()
-        self.assertEqual(n.error, "PagerTree notifications are not enabled.")
-
-    @patch("hc.api.transports.curl.request", autospec=True)
-    def test_it_does_not_escape_title(self, mock_post: Mock) -> None:
-        mock_post.return_value.status_code = 200
-
-        self.check.name = "Foo & Bar"
-        self.check.save()
-
-        self.channel.notify(self.check)
+        method, url = mock_post.call_args.args
+        self.assertIn("https://example.net", url)
+        self.assertIn(quote("!foo:example.org"), url)
+        self.assertIn("test-token", url)
 
         payload = mock_post.call_args.kwargs["json"]
-        self.assertEqual(payload["title"], "Foo & Bar is DOWN")
+        self.assertIn("Foo is DOWN.", payload["body"])
+        self.assertIn("Last ping was an hour ago.", payload["body"])
 
     @patch("hc.api.transports.curl.request", autospec=True)
     def test_it_handles_no_last_ping(self, mock_post: Mock) -> None:
         self.check.last_ping = None
         self.check.save()
-        mock_post.return_value.status_code = 200
 
+        mock_post.return_value.status_code = 200
         self.channel.notify(self.check)
+        assert Notification.objects.count() == 1
 
         payload = mock_post.call_args.kwargs["json"]
-        self.assertNotIn("Last ping was", payload["description"])
+        self.assertNotIn("Last ping was", payload["body"])